fix: 인터페이스, 구현체 네이밍 룰 획일화
