+++ app/src/main/java/com/growit/app/resource/domain/jobrole/service/JobRoleValidator.java
@@ -1,0 +1,7 @@
+package com.growit.app.resource.domain.jobrole.service;
+
+import com.growit.app.common.exception.BadRequestException;
+
+public interface JobRoleValidator {
+  void checkJobRoleExist(String id) throws BadRequestException;
+}
